% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDSC-package.R
\docType{package}
\name{FDSC-package}
\alias{FDSC}
\alias{FDSC-package}
\title{FDSC: Functional Data Subgroup Clustering and Outlier Identification via Minimum Spanning Tree}
\description{
This package implements two functional data clustering algorithms based on minimum spanning tree (MST). FDSC-MST solves the subgroup clustering problem by constructing an MST using geographical distance and functional curve similarity, and applies subgroup pairwise penalties to MST edges. An alternating direction method of multipliers (ADMM) is utilized to solve the optimization problem. For contaminated data, this package implements FDSCOD-MST, which simultaneously performs clustering and outlier detection.
}
