% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDSC-MST.R
\name{fun_est_ls_mst}
\alias{fun_est_ls_mst}
\title{ADMM algorithm
Input:}
\usage{
fun_est_ls_mst(
  fdy,
  timerange,
  B,
  A,
  diff_mat,
  mst_mat,
  gamma_old,
  vl_old,
  zeta_old,
  rho = 1,
  aa = 3,
  lambda = 0.1,
  tau = 20,
  ep = 0.1
)
}
\arguments{
\item{fdy}{a numeric matrix of dimension n * T, representing the functional data for n observations evaluated at T time points}

\item{timerange}{numeric vector of time points}

\item{B}{B=kronecker(diag(1,n),B.basis) B-spline bases, as the role of covariates in regression-based methods}

\item{A}{A=kronecker(mst_mat,diag(p))}

\item{diff_mat}{The second order difference operator}

\item{mst_mat}{(n-1)*n matrix obtained based on the constructed MST}

\item{gamma_old}{Initial value of gamma (gamma in our manuscript)}

\item{vl_old}{Initial value of vl (v_l in our manuscript)}

\item{zeta_old}{Initial value of zeta (zeta in our manuscript)}

\item{rho}{ADMM learning rate parameter rho}

\item{aa}{concavity parameter a of the MCP penalty}

\item{lambda}{the smoothing parameter lambda}

\item{tau}{the regularization parameter tau refers to MCP penalty}

\item{ep}{ADMM stopping tolerance (default = 1e-1)
Output:}
}
\value{
gamma_tmp estimated gamma by ADMM iteration (gamma in our manuscript)

vl_tmp estimated vl by ADMM iteration (v_l in our manuscript)
}
\description{
ADMM algorithm
Input:
}
