% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDSC-MST.R
\name{fun_group_mst}
\alias{fun_group_mst}
\title{Input:}
\usage{
fun_group_mst(fdy, vl_new_ls, edges)
}
\arguments{
\item{fdy}{a numeric matrix of dimension n * T, representing the functional data for n observations evaluated at T time points}

\item{vl_new_ls}{Output from function "fun_est_ls_mst" $vl.tmp}

\item{edges}{edges of construced MST
Output:}
}
\value{
group_id list of true cluster memberships

K_hat estimated number of groups
}
\description{
Input:
}
