% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDSC-class.R
\name{print.fdsc}
\alias{print.fdsc}
\title{Print method for fdsc objects}
\usage{
\method{print}{fdsc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fdsc"}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print method for fdsc objects
}
