% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDSC-class.R
\name{summary.fdsc}
\alias{summary.fdsc}
\title{Summary method for fdsc objects}
\usage{
\method{summary}{fdsc}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fdsc"}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Summary method for fdsc objects
}
