% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDSCOD-MST.R
\name{fun_statistic}
\alias{fun_statistic}
\title{fun_statistic
Input:}
\usage{
fun_statistic(timerange, fdy, out_pot, group_id_hat)
}
\arguments{
\item{timerange}{numeric vector of time points}

\item{fdy}{a numeric matrix of dimension n * T, representing the functional data for n observations evaluated at T time points}

\item{out_pot}{the potential outlying set obtained from the Preliminary Outlier Screening step in Algorithm 2}

\item{group_id_hat}{initial cluster memberships obtained from the Initial Clustering step in Algorithm 2
Output:}
}
\value{
T_stat: Test statistics for potential outliers (see equation (7) in the paper)

D: number of retained functional principal components
}
\description{
fun_statistic
Input:
}
