% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDSCOD-MST.R
\name{fun_test}
\alias{fun_test}
\title{Input:}
\usage{
fun_test(timerange, fdy, out_pot, group_id_hat, alpha0)
}
\arguments{
\item{timerange}{numeric vector of time points}

\item{fdy}{a numeric matrix of dimension n * T, representing the functional data for n observations evaluated at T time points}

\item{out_pot}{the potential outlying set obtained from the Preliminary Outlier Screening step in Algorithm 2}

\item{group_id_hat}{initial cluster memberships obtained from the Initial Clustering step in Algorithm 2}

\item{alpha0}{default 0.05 the significance level for FDR-controlled outlier detection
Output:}
}
\value{
out_set: the final detected outlier set
}
\description{
Input:
}
